Aws3_Boto3/requirements.txt
asgiref==3.3.1
boto3==1.16.25
botocore==1.19.25
Django==3.1.3
django-storages==1.10.1
djangorestframework==3.12.2
jmespath==0.10.0
Pillow==8.0.1
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2